requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
idna==2.9
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
multidict==4.7.5
pyee==7.0.1
requests==2.23.0
six==1.14.0
slackclient==2.5.0
slackeventsapi==2.1.0
SQLAlchemy==1.3.15
typing-extensions==3.7.4.1
urllib3==1.25.8
Werkzeug==1.0.0
yarl==1.4.2
